requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.0
setuptools>=50.0.0

